+++ examples/default/ios/InstabugExample/AppDelegate.mm
@@ -1,6 +1,8 @@
 #import "AppDelegate.h"
+#import <GoogleMaps/GoogleMaps.h>
 
 #import <React/RCTBundleURLProvider.h>
+#import "RNCConfig.h"
 
 @implementation AppDelegate
 
@@ -10,6 +12,9 @@
   // You can add your custom initial props in the dictionary below.
   // They will be passed down to the ViewController used by React Native.
   self.initialProps = @{};
+  NSString *googleApiKey = [RNCConfig envFor:@"GOOGLE_MAPS_API_KEY"];
+
+ [GMSServices provideAPIKey:googleApiKey]; // use the api key obtained from Google Console
 
   return [super application:application didFinishLaunchingWithOptions:launchOptions];
 }