+++ examples/default/src/App.tsx
@@ -4,6 +4,13 @@ import { StyleSheet } from 'react-native';
 import { GestureHandlerRootView } from 'react-native-gesture-handler';
 import { NavigationContainer } from '@react-navigation/native';
 import Instabug, { APM, InvocationEvent, LogLevel, ReproStepsMode } from 'instabug-reactnative';
+
+import Instabug, {
+  CrashReporting,
+  InvocationEvent,
+  LogLevel,
+  ReproStepsMode,
+} from 'instabug-reactnative';
 import { NativeBaseProvider } from 'native-base';
 
 import { RootTabNavigator } from './navigation/RootTab';
@@ -18,6 +25,7 @@ export const App: React.FC = () => {
       debugLogsLevel: LogLevel.verbose,
     });
     APM.setEnabled(true);
+    CrashReporting.setNDKCrashesEnabled(true);
 
     Instabug.setReproStepsConfig({
       all: ReproStepsMode.enabled,