+++ examples/default/src/native/NativeCrashReporting.ts
@@ -8,6 +8,20 @@ export interface CrashReportingExampleNativeModule extends NativeModule {
   sendFatalHang(): Promise<void>;
   sendANR(): Promise<void>;
   sendOOM(): Promise<void>;
+
+  sendNDKCrash(): Promise<void>;
+
+  causeSIGSEGVCrash(): Promise<void>;
+
+  causeSIGABRTCrash(): Promise<void>;
+
+  causeSIGFPECrash(): Promise<void>;
+
+  causeSIGILLCrash(): Promise<void>;
+
+  causeSIGBUSCrash(): Promise<void>;
+
+  causeSIGTRAPCrash(): Promise<void>;
 }
 
 export const NativeExampleCrashReporting = NativeExampleModules.CrashReportingExampleModule;